I init getInitialMessages function.
