+++ lib/stack/getInitialMessages.tsx
@@ -1,0 +1,31 @@
+import { Address, getAddress } from "viem";
+import { CHAT_POINT_SYSTEM_ID, MESSAGE_SENT_EVENT } from "../consts";
+import getStackClient from "./getStackClient";
+import { Message } from "ai";
+
+const getInitialMessages = async (walletAddress: Address) => {
+  const stackClient = getStackClient(CHAT_POINT_SYSTEM_ID);
+
+  const events = await stackClient.getEvents({
+    query: stackClient
+      .eventsQuery()
+      .where({
+        eventType: MESSAGE_SENT_EVENT,
+        associatedAccount: getAddress(walletAddress),
+      })
+      .offset(0)
+      .build(),
+  });
+  console.log("events", events);
+  const messages: Message[] = events.map((event) => ({
+    id: event.metadata.id,
+    content: event.metadata.content,
+    role: event.metadata.role as Message["role"],
+    createdAt: new Date(event.timestamp),
+  }));
+  console.log("messages", messages);
+  messages.sort((a, b) => a.createdAt!.getTime() - b.createdAt!.getTime());
+  return messages;
+};
+
+export default getInitialMessages;
